+++ CHANGELOG.md
@@ -1,6 +1,8 @@
-%% NEXT_VERSION_TYPE=MAJOR|MINOR|PATCH
+%% NEXT_VERSION_TYPE=MINOR
 %% NEXT_VERSION_DESCRIPTION_BEGIN
+* Добавлен функционал проверки snapshot-зависимостей
 %% NEXT_VERSION_DESCRIPTION_END
+
 ## [4.3.0]() (01-03-2019)
 
 * Переход на platformGradleProjectVersion 4 версии