+++ CHANGELOG.md
@@ -1,7 +1,7 @@
-### NEXT_VERSION_TYPE=PATCH
-### NEXT_VERSION_DESCRIPTION_BEGIN
+## [4.4.3]() (11-03-2019)
+
 * Исправление java.lang.IllegalStateException: Duplicate key 3.0.1 при выводе списка новых зависимостей
-### NEXT_VERSION_DESCRIPTION_END
+
 ## [4.4.2]() (07-03-2019)
 
 * Изменила имя таски с "checkSnapshotDependencies" на "checkSnapshotsDependencies"